Move L4 metrics related code to a separate file.
