+++ pkg/metrics/metrics.go
@@ -63,20 +63,6 @@ var (
 		},
 		[]string{label},
 	)
-	l4ILBCount = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "number_of_l4_ilbs",
-			Help: "Number of L4 ILBs",
-		},
-		[]string{label},
-	)
-	l4NetLBCount = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "number_of_l4_netlbs",
-			Help: "Number of L4 NetLBs",
-		},
-		[]string{label},
-	)
 	serviceAttachmentCount = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "number_of_service_attachments",
@@ -99,26 +85,6 @@ var (
 		[]string{"component_version"},
 	)
 )
-
-// netLBFeatureCount define metric feature count for L4NetLB controller
-type netLBFeatureCount struct {
-	service            int
-	managedStaticIP    int
-	premiumNetworkTier int
-	success            int
-	inUserError        int
-	inError            int
-}
-
-func (netlbCount *netLBFeatureCount) record() {
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBService.String()}).Set(float64(netlbCount.service))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBStaticIP.String()}).Set(float64(netlbCount.success))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBPremiumNetworkTier.String()}).Set(float64(netlbCount.premiumNetworkTier))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBManagedStaticIP.String()}).Set(float64(netlbCount.managedStaticIP))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBInSuccess.String()}).Set(float64(netlbCount.success))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBInUserError.String()}).Set(float64(netlbCount.inUserError))
-	l4NetLBCount.With(prometheus.Labels{label: l4NetLBInError.String()}).Set(float64(netlbCount.inError))
-}
 
 // init registers ingress usage metrics.
 func init() {
@@ -158,12 +124,10 @@ type ControllerMetrics struct {
 	ingressMap map[string]IngressState
 	// negMap is a map between service key to neg state
 	negMap map[string]NegServiceState
-	// l4ILBServiceMap is a map between service key and L4 ILB service state.
-	l4ILBServiceMap map[string]L4ILBServiceState
+	// l4ControllerMetrics contains data needed to calculate L4 metrics.
+	l4ControllerMetrics *l4ControllerMetrics
 	// l4ILBDualStackServiceMap is a map between service key and L4 ILB DualStack service state.
 	l4ILBDualStackServiceMap map[string]L4DualStackServiceState
-	// l4NetLBServiceMap is a map between service key and L4 NetLB service state.
-	l4NetLBServiceMap map[string]L4NetLBServiceState
 	// l4NetLBDualStackServiceMap is a map between service key and L4 NetLB DualStack service state.
 	l4NetLBDualStackServiceMap map[string]L4DualStackServiceState
 	// pscMap is a map between the service attachment key and PSC state
@@ -183,14 +147,12 @@ func NewControllerMetrics(exportInterval, l4NetLBProvisionDeadline time.Duration
 	return &ControllerMetrics{
 		ingressMap:                 make(map[string]IngressState),
 		negMap:                     make(map[string]NegServiceState),
-		l4ILBServiceMap:            make(map[string]L4ILBServiceState),
 		l4ILBDualStackServiceMap:   make(map[string]L4DualStackServiceState),
-		l4NetLBServiceMap:          make(map[string]L4NetLBServiceState),
 		l4NetLBDualStackServiceMap: make(map[string]L4DualStackServiceState),
 		pscMap:                     make(map[string]pscmetrics.PSCState),
 		serviceMap:                 make(map[string]struct{}),
 		metricsInterval:            exportInterval,
-		l4NetLBProvisionDeadline:   l4NetLBProvisionDeadline,
+		l4ControllerMetrics:        newL4ontrollerMetrics(l4NetLBProvisionDeadline),
 	}
 }
 
@@ -266,51 +228,6 @@ func (im *ControllerMetrics) DeleteNegService(svcKey string) {
 	delete(im.negMap, svcKey)
 }
 
-// SetL4ILBService implements L4ILBMetricsCollector.
-func (im *ControllerMetrics) SetL4ILBService(svcKey string, state L4ILBServiceState) {
-	im.Lock()
-	defer im.Unlock()
-
-	if im.l4ILBServiceMap == nil {
-		klog.Fatalf("Ingress Metrics failed to initialize correctly.")
-	}
-	im.l4ILBServiceMap[svcKey] = state
-}
-
-// DeleteL4ILBService implements L4ILBMetricsCollector.
-func (im *ControllerMetrics) DeleteL4ILBService(svcKey string) {
-	im.Lock()
-	defer im.Unlock()
-
-	delete(im.l4ILBServiceMap, svcKey)
-}
-
-// SetL4NetLBService adds metric state for given service to map.
-func (im *ControllerMetrics) SetL4NetLBService(svcKey string, state L4NetLBServiceState) {
-	im.Lock()
-	defer im.Unlock()
-
-	if im.l4NetLBServiceMap == nil {
-		klog.Fatalf("L4 Net LB Metrics failed to initialize correctly.")
-	}
-
-	if !state.InSuccess {
-		if previousState, ok := im.l4NetLBServiceMap[svcKey]; ok && previousState.FirstSyncErrorTime != nil {
-			// If service is in Error state and retry timestamp was set then do not update it.
-			state.FirstSyncErrorTime = previousState.FirstSyncErrorTime
-		}
-	}
-	im.l4NetLBServiceMap[svcKey] = state
-}
-
-// DeleteL4NetLBService deletes service from metrics map.
-func (im *ControllerMetrics) DeleteL4NetLBService(svcKey string) {
-	im.Lock()
-	defer im.Unlock()
-
-	delete(im.l4NetLBServiceMap, svcKey)
-}
-
 // SetServiceAttachment adds sa state to the map to be counted during metrics computation.
 // SetServiceAttachment implements PSCMetricsCollector.
 func (im *ControllerMetrics) SetServiceAttachment(saKey string, state pscmetrics.PSCState) {
@@ -372,17 +289,8 @@ func (im *ControllerMetrics) export() {
 	}
 	klog.V(3).Infof("Ingress usage metrics exported.")
 
-	ilbCount := im.computeL4ILBMetrics()
-	klog.V(3).Infof("Exporting L4 ILB usage metrics: %#v", ilbCount)
-	for feature, count := range ilbCount {
-		l4ILBCount.With(prometheus.Labels{label: feature.String()}).Set(float64(count))
-	}
-	klog.V(3).Infof("L4 ILB usage metrics exported.")
-
-	netlbCount := im.computeL4NetLBMetrics()
-	klog.V(3).Infof("Exporting L4 NetLB usage metrics: %#v", netlbCount)
-	netlbCount.record()
-	klog.V(3).Infof("L4 NetLB usage metrics exported.")
+	// Export L4 metrics.
+	im.l4ControllerMetrics.export()
 
 	im.exportL4ILBDualStackMetrics()
 	im.exportL4NetLBDualStackMetrics()
@@ -498,78 +406,6 @@ func (im *ControllerMetrics) computeNegMetrics() map[feature]int {
 		}
 	}
 	klog.V(4).Info("NEG usage metrics computed.")
-	return counts
-}
-
-// computeL4ILBMetrics aggregates L4 ILB metrics in the cache.
-func (im *ControllerMetrics) computeL4ILBMetrics() map[feature]int {
-	im.Lock()
-	defer im.Unlock()
-	klog.V(4).Infof("Computing L4 ILB usage metrics from service state map: %#v", im.l4ILBServiceMap)
-	counts := map[feature]int{
-		l4ILBService:      0,
-		l4ILBGlobalAccess: 0,
-		l4ILBCustomSubnet: 0,
-		l4ILBInSuccess:    0,
-		l4ILBInError:      0,
-		l4ILBInUserError:  0,
-	}
-
-	for key, state := range im.l4ILBServiceMap {
-		klog.V(6).Infof("ILB Service %s has EnabledGlobalAccess: %t, EnabledCustomSubnet: %t, InSuccess: %t", key, state.EnabledGlobalAccess, state.EnabledCustomSubnet, state.InSuccess)
-		counts[l4ILBService]++
-		if !state.InSuccess {
-			if state.IsUserError {
-				counts[l4ILBInUserError]++
-			} else {
-				counts[l4ILBInError]++
-			}
-			// Skip counting other features if the service is in error state.
-			continue
-		}
-		counts[l4ILBInSuccess]++
-		if state.EnabledGlobalAccess {
-			counts[l4ILBGlobalAccess]++
-		}
-		if state.EnabledCustomSubnet {
-			counts[l4ILBCustomSubnet]++
-		}
-	}
-	klog.V(4).Info("L4 ILB usage metrics computed.")
-	return counts
-}
-
-// computeL4NetLBMetrics aggregates L4 NetLB metrics in the cache.
-func (im *ControllerMetrics) computeL4NetLBMetrics() netLBFeatureCount {
-	im.Lock()
-	defer im.Unlock()
-	klog.V(4).Infof("Computing L4 NetLB usage metrics from service state map: %#v", im.l4NetLBServiceMap)
-	var counts netLBFeatureCount
-
-	for key, state := range im.l4NetLBServiceMap {
-		klog.V(6).Infof("NetLB Service %s has metrics %+v", key, state)
-		counts.service++
-		if state.IsUserError {
-			counts.inUserError++
-			// Skip counting other features if the service is in error state.
-			continue
-		}
-		if !state.InSuccess {
-			if time.Since(*state.FirstSyncErrorTime) >= im.l4NetLBProvisionDeadline {
-				counts.inError++
-			}
-			// Skip counting other features if the service is in error state.
-			continue
-		}
-		counts.success++
-		if state.IsManagedIP {
-			counts.managedStaticIP++
-		}
-		if state.IsPremiumTier {
-			counts.premiumNetworkTier++
-		}
-	}
-	klog.V(4).Info("L4 NetLB usage metrics computed.")
 	return counts
 }
